Create Shotgun Node.lsl
This is the shotgun node!
We are using PRIM_TEXT to transfer a base64 encoded series of integers in a strided list. We are using 3 base64s for XYZ which returns a uniform length.
As of right now, the maximum number of shots we can transfer is 23, which is honestly overkill, but it's better to have more than enough room than riding the storage space to the line.